requirements.txt
backtrader==1.9.78.123
pandas==2.0.3
numpy==1.24.3
tushare==1.2.89
typer[all]==0.9.0
matplotlib==3.7.2
seaborn==0.12.2
rich==13.5.2
python-dotenv==1.0.0
click==8.1.7 
gmapi>=1.2.0 